app/proxier/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
python-multipart